Fix a nasty bug when calling _start
This bug made i686 machines in Linux fail to start at all since all
registers where the parameters to _start are located are clobbered by
the libinteractive initialization. Fortunately, in x86_64 it just
happened that _most_ things were unchanged, so most problems would not
notice the change. The exception is that the values of argc and argv
were wrong since the stack was not perfectly aligned to what _start
expects. This change makes __entry behave _exactly_ as if nothing had
happened, so _start runs OK in all Linux platforms and configurations.