Use one pipe pair per interface
Since pipe messages from the children to Main were multiplexed, when
there was more than one child process, and one of them died, Main would
not get notified of its death (since the pipe still had at least one
process writing to it) and would hang.

This change removes the nasty hack of comparing the Main interface name
to "Main" and treating it specially, so now all non-Main interfaces have
a named pipe pair to communicate with Main. This fixes #12.