Add the [NoReturn] attribute to all Main functions
Since most of the time the functions in the Main interface perform
parameter validation, exiting in case something went wrong, all
functions in this interface should have the [NoReturn] attribute. Since
it is too onerous to require problemsetters to manually add it
themselves, make the compiler insert it for them.