Better exit status in case of errors
Most places where there is an explicit exit now uses unique exit codes
to distinguish between them. In languages that support exceptions, most
of them will continue to be used, except for error code 242, which is
useful to always handle the same way. The defined exit codes are as
follows:

* 237 - transact out of shared memory
* 238 - transact invalid state
* 239 - incomplete message
* 240 - invalid cookie
* 241 - invalid message id
* 242 - confused about exiting (other process died while expecting a
  reply)
* 243 - initialization error