Support for ARM
This change splits off the platform-specific libinteractive
initialization code (in assembly) into a separate `.S` file to avoid all
gcc-added magic to that function and adds support for ARM while we're at
it. In particular, by moving it to a separate file, we avoid
stack-modifying prologue that has been changed across gcc versions.

Fixed #22.