Add an option to specify the directory where the pipes are located
This is to make it easier to run in the Linux subsystem for Windows,
since it does not allow to add pipes in any directory, only those
that are mounted on a Linux filesystem (e.g. ext4 or tmpfs).

Using this setting on real Windows is not supported, since
CreateNamedPipe() creates the pipe in a completely different namespace
that does not support directory creation.

Fixes: #25